Send all actions from api to redux
