+++ src/Wrido.Web/src/actionCreators.js
@@ -1,8 +1,10 @@
-import { onInputChange, } from './constants';
+import { onInputChange, queryCancelled, queryCompleted, queryExecuting, queryReceived, resultsAvailable } from './constants';
 
-export const onInputChangeAction = value => ({
-  type: onInputChange,
-  payload: {
-    value
-  }
-});
+const createActionCreator = (type, payloader) => payload => ({ type, payload: payloader(payload) });
+
+export const onInputChangeAction = createActionCreator(onInputChange, value => ({ value }));
+export const queryCancelledAction = createActionCreator(queryCancelled, value => ({ value }));
+export const queryCompletedAction = createActionCreator(queryCompleted, value => ({ value }));
+export const queryExecutingAction = createActionCreator(queryExecuting, value => ({ value }));
+export const queryReceivedAction = createActionCreator(queryReceived, value => ({ value }));
+export const resultsAvailableAction = createActionCreator(resultsAvailable, value => ({ value }));
