+++ src/Wrido.Web/src/store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { rootReducer } from './modules/root';
 import { createEpicMiddleware } from 'redux-observable';
 import { rootEpic } from './modules/root';
-import { connectToSignalR } from './middlewares';
+import { connectToSignalR } from './connectToSignalR';
 
 const middleware = [
   createEpicMiddleware(rootEpic),
